extract class Framework from AndrolibResources
